Fixed player movement, added hoop features
Player is now able to shoot to the hoop and hoop is able to tell if the player is scored.

When shooting, now player will to to the right. To be fixed as the hoops direction.

Player now shoots when Spacebar is released, not when the maximum jumping limit reached.

Ligting is updated.